server/requirements.txt
python-dotenv
pycryptodome
pyodbc
requests